[dsa] Extract memory_representation into the dsa library
This functionally does very similar things to the code we have in
visual. We are refactoring it so that the code is shared.
